refactor: fix components import/export
