+++ tsconfig.json
@@ -22,7 +22,8 @@
 		"baseUrl": "./src",
 		"paths": {
 			"@api": ["./api/index.ts"],
-			"@components": ["./components/index.ts"],
+			"@components/layout": ["./components/layout/index.ts"],
+			"@components/ui": ["./components/ui/index.ts"],
 			"@assets": ["./assets/index.ts"]
 		}
 	},
